mlb-data-analysis/requirements.txt
MLB-StatsAPI
pybaseball
pytest >= 8